requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
PyMuPDF
sentence-transformers
faiss-cpu